fixed a couple of bugs in the code that repairs the systemd service. Also, added a check to make sure that the old broken service exists before we try to fix it.

+++ libraries/helpers_centos.rb
@@ -5,19 +5,19 @@ module ClamavCookbook
       def centos7_post_install
         proper_server_service = '/usr/lib/systemd/system/clamd@server.service'
         broken_server_service = '/usr/lib/systemd/system/clamd@.service'
-        file server_service do
-          owner 'root'
-          group 'root'
-          mode 0755
-          content ::File.open(broken_server_service).read
-          action :create
-          not_if { File.exist? server_service }
-        end
-        file broken_server_service do
-          action :delete
+        if File.exist? broken_server_service
+          file proper_server_service do
+            owner 'root'
+            group 'root'
+            mode 0755
+            content ::File.open(broken_server_service).read
+            action :create_if_missing
+          end
+          file broken_server_service do
+            action :delete
+          end
         end
       end
     end
   end
 end
-
